scripts/igep-media-create: copy README file to SD-card if exist.
When the script generates a SD-card the rootfs image is copied to
the /opt/firmware directory, as would be interesting have the README
file into the SD-card this patch adds the file to the /opt/firmware.

This script can be used outside a firmware directory passing the image
file, so it's possible that the README doesn't exist, in that case simply
notice that the README file is not found and is not being copied to the
SD-card, and continue without error.

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>
Acked-by: Matthias Brugger <mbrugger@iseebcn.com>